The original paper and my correction
Turing was one of the greatest minds of all time. I am not working to surpass his contributions, only to rectify an honest, very human omission in his work that was a result of the zeitgeist of his time. Gödel's incompleteness was his inspiration for his work on the Entscheidungsproblem, and mathematicians had a surge of breakthroughs by imposing arbitrary limitations on their thinking. However, I believe a blind spot was created, which has been perpetuated to this day. The only way science has any leigitmacy at all is if we have the ability to approach out legends and heroes as mere men of mortality like ourselves and not as the gods of which they have been portrayed. The only requirement is an open mind and years of study.