Added draft of earlier version of proof
As we know, Turing's proof on the Entscheidungs problem reduces to Cantor's Diagonalization. If there is a method to find exceptions in one, there should be a method to find exceptions in the other, independently, through a different method. This paper was written before the paper I wrote on Turing. I wrote the paper on Turing in response to those who has trouble following or believing the exception in Cantor's argument. I believe that the fact I have found two, independent methods, each of which provides exceptions to incomplete proofs that reduce to each other, is not to be taken lightly.

It says "final" in the title of the file, but maybe in 10 years, i'll revisit it, and change the notation and description of the diagonalization function.